-Added Action class with methods
