+++ Main/src/Coffee/MachinesHaveNeeds.java
@@ -1,0 +1,31 @@
+package Coffee;
+
+import java.util.Scanner;
+
+public class MachinesHaveNeeds {
+
+    // sMachinesHaveNeeds.setIngredients(MachinesHaveNeeds.getCups()); // psvm part to run homework 2
+
+    public static int getCups() {
+        Scanner in = new Scanner(System.in);
+        System.out.print("Write how many cups of coffee you will need...");
+        int cups = in.nextInt();
+        System.out.println("For " + cups + " cups of coffee you will need:");
+        in.close();
+        return cups;
+    }
+    public static void setIngredients(int cups) {
+        int water = 200;
+        int milk = 50;
+        int coffeeBeans = 15;
+        water = cups * water;
+        milk = cups * milk;
+        coffeeBeans = cups * coffeeBeans;
+        getIngredients(water, milk, coffeeBeans);
+    }
+    public static void getIngredients(int water, int milk, int coffeeBeans){
+        System.out.println(water + " ml of water");
+        System.out.println(milk + " ml of milk");
+        System.out.println(coffeeBeans + " gr of coffee beans");
+    }
+}
